src/main/java/com/training/springcore/service/SiteServiceImpl.java
package com.training.springcore.service;

import com.training.springcore.model.Site;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.context.annotation.Scope;
import org.springframework.core.io.Resource;
import org.springframework.core.io.ResourceLoader;
import org.springframework.stereotype.Service;

import java.io.IOException;
import java.io.InputStream;
import java.util.Scanner;


@Service
@Scope("prototype")
public class SiteServiceImpl implements SiteService {

    @Autowired
    private ResourceLoader resourceLoader;

    private CaptorService captorService;

    public SiteServiceImpl() {
    }

    @Autowired
    public SiteServiceImpl(CaptorService captorService) {
        System.out.println("1. Init SiteServiceImpl :::   " + this);
        this.captorService = captorService;
    }

    @Override
    public Site findById(String siteId) {
        System.out.println("2. Appel de findById :::  " + this);
        if (siteId == null) {
            return null;
        }

        Site site = new Site("Florange");
        site.setId(siteId);
        site.setCaptors(captorService.findBySite(siteId));
        return site;
    }

    @Override
    public void readFile(String path) {


        Resource resource = resourceLoader
                .getResource(path);
        try (InputStream stream = resource.getInputStream()) {
            Scanner scanner = new Scanner(stream).useDelimiter("\\n");
            while (scanner.hasNext()) {
                System.out.println(scanner.next());
            }
        }
        catch (IOException e) {
            e.printStackTrace();
        }

    }
}
